Fixed incomplete .obj rendering problem

+++ OpenGLScene/OpenGLScene/LoadFunctions.cpp
@@ -349,7 +349,7 @@ bool LoadOBJ(  const char* obj_file_path
 
 bool LoadAssImp(
                 const char * path,
-                std::vector<unsigned short> & indices,
+                std::vector<unsigned int> & indices,
                 std::vector<glm::vec3> & vertices,
                 std::vector<glm::vec2> & uvs,
                 std::vector<glm::vec3> & normals
