+++ OpenGLScene/OpenGLScene/LoadFunctions.hpp
@@ -43,7 +43,7 @@ bool LoadOBJ(
              );
 bool LoadAssImp(
                 const char * path,
-                std::vector<unsigned short> & indices,
+                std::vector<unsigned int> & indices,
                 std::vector<glm::vec3> & vertices,
                 std::vector<glm::vec2> & uvs,
                 std::vector<glm::vec3> & normals
